updated architecture
