+++ Cargo.toml
@@ -54,9 +54,12 @@ wasi-common = { workspace = true, default-features = true, features = ["exit"], 
 wasmtime-wasi = { workspace = true, default-features = true, optional = true }
 wasmtime-wasi-nn = { workspace = true, optional = true }
 wasmtime-wasi-threads = { workspace = true, optional = true }
-wasmtime-lind-fork = { path = "crates/lind-fork" }
+wasmtime-lind = { path = "crates/lind" }
+wasmtime-lind-common = { path = "crates/lind-common" }
+wasmtime-lind-utils = { path = "crates/lind-utils" }
 wasmtime-wasi-http = { workspace = true, optional = true }
 rustposix = { workspace = true }
+rawposix = { workspace = true }
 clap = { workspace = true }
 anyhow = { workspace = true, features = ['std'] }
 target-lexicon = { workspace = true }
@@ -150,6 +153,7 @@ members = [
   "fuzz",
   "winch/codegen",
   "crates/rustposix",
+  "crates/rawposix",
 ]
 exclude = [
   'docs/rust_wasi_markdown_parser',
@@ -196,7 +200,9 @@ wasmtime-wasi = { path = "crates/wasi", version = "23.0.0", default-features = f
 wasmtime-wasi-http = { path = "crates/wasi-http", version = "=23.0.0", default-features = false }
 wasmtime-wasi-nn = { path = "crates/wasi-nn", version = "23.0.0" }
 wasmtime-wasi-threads = { path = "crates/wasi-threads", version = "23.0.0" }
-wasmtime-lind-fork = { path = "crates/lind-fork", version = "23.0.0" }
+wasmtime-lind = { path = "crates/lind", version = "23.0.0" }
+wasmtime-lind-common = { path = "crates/lind-common", version = "23.0.0" }
+wasmtime-lind-utils = { path = "crates/lind-utils", version = "23.0.0" }
 wasmtime-component-util = { path = "crates/component-util", version = "=23.0.0" }
 wasmtime-component-macro = { path = "crates/component-macro", version = "=23.0.0" }
 wasmtime-asm-macros = { path = "crates/asm-macros", version = "=23.0.0" }
@@ -213,6 +219,7 @@ wasmtime-jit-icache-coherence = { path = "crates/jit-icache-coherence", version 
 wasmtime-wit-bindgen = { path = "crates/wit-bindgen", version = "=23.0.0" }
 test-programs-artifacts = { path = 'crates/test-programs/artifacts' }
 rustposix = { path = "crates/rustposix" }
+rawposix = { path = "crates/rawposix" }
 
 cranelift-wasm = { path = "cranelift/wasm", version = "0.110.0" }
 cranelift-codegen = { path = "cranelift/codegen", version = "0.110.0", default-features = false, features = ["std", "unwind", "trace-log"] }
