--- crates/lind-fork/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "wasmtime-lind-fork"
-version.workspace = true
-authors.workspace = true
-description = "lind fork implementation for wasmtime"
-edition.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-anyhow = { workspace = true }
-log = { workspace = true }
-rand = "0.8"
-wasi-common = { workspace = true, features = ["exit"]}
-wasmtime = { workspace = true, features = ['threads'] }
-wasmtime-environ = { workspace = true }
-rustposix = { path = "../rustposix" }
