+++ crates/lind-utils/src/lib.rs
@@ -1,0 +1,40 @@
+#![allow(dead_code)]
+
+use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::{Arc, Barrier, Condvar, Mutex};
+
+#[allow(missing_docs)]
+#[derive(Default)]
+pub struct LindCageManager {
+    cage_count: Mutex<i32>,
+    condvar: Condvar,
+}
+
+impl LindCageManager {
+    pub fn new(value: i32) -> Self {
+        LindCageManager {
+            cage_count: Mutex::new(value),
+            condvar: Condvar::new(),
+        }
+    }
+
+    pub fn increment(&self) {
+        let mut cage_count = self.cage_count.lock().unwrap();
+        *cage_count += 1;
+    }
+
+    pub fn decrement(&self) {
+        let mut cage_count = self.cage_count.lock().unwrap();
+        *cage_count -= 1;
+        if *cage_count == 0 {
+            self.condvar.notify_all();
+        }
+    }
+
+    pub fn wait(&self) {
+        let mut cage_count = self.cage_count.lock().unwrap();
+        while *cage_count != 0 {
+            cage_count = self.condvar.wait(cage_count).unwrap();
+        }
+    }
+}
